+++ src/shared/chat/styles/message-body.scss
@@ -1,0 +1,5 @@
+converse-chat-message-body {
+    audio {
+        width: 100%;
+    }
+}
