+++ src/shared/components/rich-text.js
@@ -6,6 +6,8 @@ export default class RichText extends CustomElement {
 
     static get properties () {
         return {
+            embed_audio: { type: Boolean },
+            embed_videos: { type: Boolean },
             mentions: { type: Array },
             nick: { type: String },
             offset: { type: Number },
@@ -13,7 +15,6 @@ export default class RichText extends CustomElement {
             onImgLoad: { type: Function },
             render_styling: { type: Boolean },
             show_images: { type: Boolean },
-            embed_videos: { type: Boolean },
             show_me_message: { type: Boolean },
             text: { type: String },
         }
@@ -21,22 +22,24 @@ export default class RichText extends CustomElement {
 
     constructor () {
         super();
-        this.offset = 0;
+        this.embed_audio = false;
+        this.embed_videos = false;
         this.mentions = [];
+        this.offset = 0;
         this.render_styling = false;
         this.show_images = false;
-        this.embed_videos = false;
         this.show_me_message = false;
     }
 
     render () {
         const options = {
+            embed_audio: this.embed_audio,
+            embed_videos: this.embed_videos,
             nick: this.nick,
             onImgClick: this.onImgClick,
             onImgLoad: this.onImgLoad,
             render_styling: this.render_styling,
             show_images: this.show_images,
-            embed_videos: this.embed_videos,
             show_me_message: this.show_me_message,
         }
         return renderRichText(this.text, this.offset, this.mentions, options);
